refactor: 불필요한 디렉토리 삭제 및 이동
