+++ koget.xcodeproj/xcshareddata/xcschemes/LockScreenWidgetExtension.xcscheme
@@ -57,8 +57,18 @@
       debugServiceExtension = "internal"
       allowLocationSimulation = "YES"
       launchAutomaticallySubstyle = "2">
-      <BuildableProductRunnable
-         runnableDebuggingMode = "0">
+      <RemoteRunnable
+         runnableDebuggingMode = "2"
+         BundleIdentifier = "com.apple.springboard">
+         <BuildableReference
+            BuildableIdentifier = "primary"
+            BlueprintIdentifier = "3AF9F00328ED570D00F685E6"
+            BuildableName = "LockScreenWidgetExtension.appex"
+            BlueprintName = "LockScreenWidgetExtension"
+            ReferencedContainer = "container:koget.xcodeproj">
+         </BuildableReference>
+      </RemoteRunnable>
+      <MacroExpansion>
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "3ABA6C1228C6043F005AA8E4"
@@ -66,7 +76,7 @@
             BlueprintName = "koget"
             ReferencedContainer = "container:koget.xcodeproj">
          </BuildableReference>
-      </BuildableProductRunnable>
+      </MacroExpansion>
       <EnvironmentVariables>
          <EnvironmentVariable
             key = "_XCWidgetKind"
