--- koget/Helper/Constant/WidgetConstants.swift
@@ -1,12 +1,0 @@
-//
-//  WidgetConstants.swift
-//  koget
-//
-//  Created by Heonjin Ha on 2023/02/16.
-//
-
-import Foundation
-
-//let idSeparator = "Z8a~du!!"
-//let schemeToAppLink = "link://"
-//let schemeToOpenLink = "koget://"
