+++ koget/KogetApp.swift
@@ -50,15 +50,22 @@ struct KogetApp: App {
     /// 위젯 scheme을 확인하고 deepLink를 엽니다.
     func maybeOpenedFromWidget(urlString: String) {
         // print("‼️위젯으로 앱을 열었습니다. ")
-        let separatedURL = urlString.split(separator: idSeparator, maxSplits: 1)
+        let separatedURL = urlString.split(separator: WidgetConstant.idSeparator, maxSplits: 1)
         let url = String(separatedURL[0]).deletingPrefix(WidgetConstant.mainURL)
         let id = String(separatedURL[1])
         
         coreData.linkWidgets.contains { deepLink in
             if deepLink.id?.uuidString == id {
                 deepLink.runCount += 1
-                coreData.saveData()
-                coreData.loadData()
+                coreData.saveData { error in
+                    if let error = error {
+                        print(error.localizedDescription)
+                        return
+                    }
+
+                    WidgetCoreData.shared.loadData()
+
+                }
                 return true
             } else {
                 return false
