+++ koget/Services/CoreData/WidgetModel.xcdatamodeld/WidgetModel.xcdatamodel/contents
