+++ koget/View/Detail/DetailWidgetView.swift
@@ -114,7 +114,15 @@ struct DetailWidgetView: View {
         if selectedWidget.opacity == nil {
             selectedWidget.opacity = 1.0
             viewModel.opacityValue = 1.0
-            coreData.saveData()
+            coreData.saveData { error in
+                if let error = error {
+                    print(error.localizedDescription)
+                    return
+                }
+
+                WidgetCoreData.shared.loadData()
+
+            }
         } else {
             viewModel.opacityValue = selectedWidget.opacity as? Double ?? 1.0
         }
